do not install to core, starting with perl v5.12.0
For one thing, v5.12.0 fixed the ordering of @INC so that site comes
before core and things installed to site from the CPAN will be
correctly loaded before the version from core is found.

For another, the install destination will need to be site to make the
CPAN-installed Pod::LaTeX not trigger deprecation warnings!